+++ campus-property-ui/pom.xml
@@ -54,20 +54,6 @@
                             </arguments>
                         </configuration>
                     </execution>
-                    <execution>
-                        <id>npm run clean (clean)</id>
-                        <goals>
-                            <goal>exec</goal>
-                        </goals>
-                        <phase>clean</phase>
-                        <configuration>
-                            <executable>npm</executable>
-                            <arguments>
-                                <argument>run</argument>
-                                <argument>clean</argument>
-                            </arguments>
-                        </configuration>
-                    </execution>
                 </executions>
             </plugin>
         </plugins>
