+++ campus-property-ui/src/App.js
@@ -1,19 +1,14 @@
-import React, {Component} from "react";
-import axios from "axios";
-import TestTable from "./campus/TestTable";
+import React, { Component } from "react";
+import RequestListForm from "./campus/RequestListForm";
 
 class App extends Component {
   state = {
     text: "initialText"
   };
 
-  componentDidMount() {
-    axios.get("/hello").then(({ data }) => this.setState({ text: data }));
-  }
-
   render() {
     return (
-      <TestTable title={this.state.text}/>
+      <RequestListForm />
     );
   }
 }
